+++ packages/workspace-fusion/src/lib/integrations/data-source/utils/prepareConfig.ts
@@ -2,10 +2,11 @@ import { UseQueryOptions } from '@tanstack/react-query';
 import { createFetchFunction, DataSourceConfig } from '..';
 
 export function prepareConfig(
-	config?: DataSourceConfig<Record<PropertyKey, unknown>>
+	config: DataSourceConfig<Record<PropertyKey, unknown>> | undefined,
+	appKey: string
 ): UseQueryOptions<unknown, unknown> {
 	return {
-		queryKey: config?.queryKey ?? ['Workspace'],
+		queryKey: config?.queryKey ?? ['Workspace', appKey],
 		initialData: config?.initialData,
 		retry: 0,
 		useErrorBoundary: true,