+++ packages/workspace-fusion/src/lib/integrations/power-bi/utils/addPowerBi.tsx
@@ -3,28 +3,17 @@ import { Tab } from '@equinor/workspace-react';
 import { PowerBiHeader } from '../components/workspaceHeader/PowerBiHeader';
 import { PowerBiIcon } from '../icons/PowerBiIcon';
 import { FilterConfig, PowerBiConfig } from '../';
-import { BaseEvent } from '@equinor/workspace-core';
+import { HeaderIcon, useWorkspaceHeaderComponents } from '../../../context';
+import { useEffect } from 'react';
+import { PowerBiPopover } from '../components/PowerBiPopover';
 
-export function addPowerBi<
-	TData extends Record<PropertyKey, unknown>,
-	TError,
-	TContext extends Record<PropertyKey, unknown> = never,
-	TCustomSidesheetEvents extends BaseEvent = never
->(powerBiConfig: PowerBiConfig | undefined): undefined | Tab {
+export function addPowerBi(powerBiConfig: PowerBiConfig | undefined): undefined | Tab {
 	if (!powerBiConfig) return;
 
 	const controller = new PowerBiController();
 
 	return {
-		Component: () => (
-			<PowerBI
-				controller={controller}
-				getToken={powerBiConfig.getToken}
-				getEmbedInfo={powerBiConfig.getEmbed}
-				reportUri={powerBiConfig.reportUri}
-				filters={createBasicFilter(powerBiConfig.filters)}
-			/>
-		),
+		Component: () => <PowerBiWrapper {...powerBiConfig} controller={controller} />,
 		CustomHeader: () => <PowerBiHeader controller={controller} />,
 		name: 'powerbi',
 		TabIcon: () => <PowerBiIcon />,
@@ -41,3 +30,42 @@ function createBasicFilter(filters: FilterConfig | undefined): undefined | IBasi
 		values: filters.values,
 	};
 }
+
+const PowerBiWrapper = (powerBiConfig: PowerBiConfig & { controller: PowerBiController }) => {
+	const { setIcons } = useWorkspaceHeaderComponents();
+
+	useEffect(() => {
+		if (powerBiConfig.ReportMetaData) {
+			const { ReportMetaData } = powerBiConfig;
+
+			const icon: HeaderIcon = {
+				Icon: ({ anchor }) => (
+					<PowerBiPopover
+						reportUri={powerBiConfig.reportUri}
+						anchor={anchor}
+						ReportMetaData={ReportMetaData}
+					/>
+				),
+				name: 'report_metadata',
+				placement: 'left',
+			};
+
+			setIcons((icons) => [...icons, icon]);
+			return () => {
+				setIcons((i) => i.filter((s) => s.name !== 'report_metadata'));
+			};
+		}
+		return;
+	}, [powerBiConfig.ReportMetaData]);
+
+	return (
+		<PowerBI
+			getErrorMessage={powerBiConfig.getErrorMessage}
+			controller={powerBiConfig.controller}
+			getToken={powerBiConfig.getToken}
+			getEmbedInfo={powerBiConfig.getEmbed}
+			reportUri={powerBiConfig.reportUri}
+			filters={createBasicFilter(powerBiConfig.filters)}
+		/>
+	);
+};