+++ packages/workspace-fusion/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.2](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.1...@equinor/workspace-fusion@0.16.2) (2023-02-02)
+
+### Bug Fixes
+
+-   :bug: fix powerbi popover close ([#299](https://github.com/equinor/fusion-workspace/issues/299)) ([bebb55e](https://github.com/equinor/fusion-workspace/commit/bebb55efa245782b4ccd0a42d38e52ce2b642a20))
+
+## [0.16.1](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.0...@equinor/workspace-fusion@0.16.1) (2023-02-02)
+
+### Bug Fixes
+
+-   :bug: fix powerbi popover ([83c2129](https://github.com/equinor/fusion-workspace/commit/83c21292bffe156c479fc14045aa8577a2fd861f))
+
+## [0.16.0](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.15.4...@equinor/workspace-fusion@0.16.0) (2023-02-02)
+
+### Features
+
+-   :sparkles: add popover to powerbi ([#293](https://github.com/equinor/fusion-workspace/issues/293)) ([de5de6a](https://github.com/equinor/fusion-workspace/commit/de5de6a1032fcec2d6e72e1604eb1301d31d82ae))
+
+## [0.15.4](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.15.3...@equinor/workspace-fusion@0.15.4) (2023-02-01)
+
+### Bug Fixes
+
+-   :bug: fix no data splash screen not showing ([#292](https://github.com/equinor/fusion-workspace/issues/292)) ([38f7df5](https://github.com/equinor/fusion-workspace/commit/38f7df57a70aee1f171f809340b3ec27896b093c))
+
+## [0.15.3](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.15.2...@equinor/workspace-fusion@0.15.3) (2023-02-01)
+
+**Note:** Version bump only for package @equinor/workspace-fusion
+
 ## [0.15.2](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.15.1...@equinor/workspace-fusion@0.15.2) (2023-01-30)
 
 **Note:** Version bump only for package @equinor/workspace-fusion